+++ src/Storage.hpp
@@ -154,6 +154,8 @@ public:
     void MoveBank(int from, int to);
     int64_t DeleteBank(int64_t bankId);
 
+    bool IsInUploadsDirectory(const std::filesystem::path&path) const;
+    
     FileRequestResult GetFileList2(const std::string&relativePath,const UiFileProperty&fileProperty);
 
     FileRequestResult GetModFileList2(const std::string &relativePath,const UiFileProperty &fileProperty);